dom/payments/test/test_update_errors.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1435157
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1435157</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="DefaultData.js"></script>
  <script type="application/javascript">

  "use strict";
  SimpleTest.waitForExplicitFinish();

  var gUrl = SimpleTest.getTestFileURL('UpdateErrorsChromeScript.js');
  var gScript = SpecialPowers.loadChromeScript(gUrl);

  function testFailHandler(message) {
    ok(false, message);
  }
  function testPassHandler(message) {
    ok(true, message);
  }
  gScript.addMessageListener("test-fail", testFailHandler);
  gScript.addMessageListener("test-pass", testPassHandler);

  async function requestChromeAction(action, params) {
    await new Promise(resolve => {
      gScript.addMessageListener(`${action}-complete`, function completeListener() {
        gScript.removeMessageListener(`${action}-complete`, completeListener);
        resolve();
      });
      gScript.sendAsyncMessage(action, params);
    });
  }

  const addressErrors = {
    addressLine: "addressLine error",
    city: "city error",
    country: "country error",
    dependentLocality: "dependentLocality error",
    organization: "organization error",
    phone: "phone error",
    postalCode: "postalCode error",
    recipient: "recipient error",
    region: "region error",
    regionCode: "regionCode error",
    sortingCode: "sortingCode error",
  };

  const payErrors = {
    email: "email error",
    name: "name error",
    phone: "phone error",
  };

  let updateDetails = {
    total:{
      label: "Total",
      amount: {
        currency: "USD",
        value: "1.00",
      },
    },
    error: "shipping address error",
    shippingAddressErrors: addressErrors,
    payerErrors: payErrors,
  }

  // testing functions
  async function testUpdateErrors() {
    const testName = "testUpdateErrors";
    await requestChromeAction("setup-update-with-errors", testName);
    return new Promise(async (resolve) => {
      const request = new PaymentRequest(defaultMethods, defaultDetails, defaultOptions);
      await request.addEventListener("shippingaddresschange", event => {
        event.updateWith(updateDetails);
      });
      await request.addEventListener("shippingoptionchange", event => {
        event.updateWith(updateDetails);
      });
      const handler = SpecialPowers.getDOMWindowUtils(window).setHandlingUserInput(true);
      try {
        const response = await request.show();
        ok(false, `${testName}: expected 'AbortError', but got resolved.`);
      } catch (error) {
        is(error.name, "AbortError", `${testName}: expected 'AbortError'.`);
      }
      await handler.destruct();
      resolve();
    });
  }

  async function testNoEventHandler(testName, eventName) {
    await requestChromeAction(`setup-no-${eventName}`, testName);
    return new Promise(async (resolve) => {
      const request = new PaymentRequest(defaultMethods, defaultDetails, defaultOptions);
      const handler = SpecialPowers.getDOMWindowUtils(window).setHandlingUserInput(true);
      try {
        const response = await request.show();
        ok(false, `${testName}: expected 'AbortError', but got resolved.`);
      } catch (error) {
        is(error.name, "AbortError", `${testName}: expected 'AbortError'.`);
      }
      await handler.destruct();
      resolve();
    });
  }

  async function testNoOnshippingaddresschange() {
    const testName = "testNoOnshippingaddresschange";
    return testNoEventHandler(testName, "onshippingaddresschange");
  }

  async function testNoOnshippingoptionchange() {
    const testName = "testNoOnshippingoptionchange";
    return testNoEventHandler(testName, "onshippingoptionchange");
  }

  async function pendingPromise() {
    return new Promise((resolve) => {
    });
  }

  async function testShowWithPendingPromise() {
    const testName = "testShowWithPendingPromise";
    await requestChromeAction("setup-show-wiht-pending-promise", testName);
    return new Promise(async (resolve) => {
      const request = new PaymentRequest(defaultMethods, defaultDetails, defaultOptions);
      const handler = SpecialPowers.getDOMWindowUtils(window).setHandlingUserInput(true);
      try {
        const response = await request.show(pendingPromise());
        ok(false, `${testName}: expected 'AbortError', but got resolved.`);
      } catch (error) {
        is(error.name, "AbortError", `${testName}: expected 'AbortError'.`);
      }
      await handler.destruct();
      resolve();
    });
  }

  async function testEventHandlingTimeout() {
    const testName = "testEventHandlingTimeout";
    await requestChromeAction("setup-timeout-handling", testName);
    return new Promise(async (resolve) => {
      const request = new PaymentRequest(defaultMethods, defaultDetails, defaultOptions);
      request.addEventListener("shippingaddresschange", (event) => {});
      request.addEventListener("shippingoptoinchange", (event) => {});
      const handler = SpecialPowers.getDOMWindowUtils(window).setHandlingUserInput(true);
      try {
        const response = await request.show();
        ok(false, `${testName}: expected 'AbortError', but got resolved.`);
      } catch (error) {
        is(error.name, "AbortError", `${testName}: expected 'AbortError'.`);
      }
      await handler.destruct();
      resolve();
    });
  }

  // teardown function
  async function teardown() {
    gScript.addMessageListener("teardown-complete", function teardownCompleteHandler() {
      gScript.removeMessageListener("teardown-complete", teardownCompleteHandler);
      gScript.removeMessageListener("test-fail", testFailHandler);
      gScript.removeMessageListener("test-pass", testPassHandler);
      gScript.destroy();
      SimpleTest.finish();
    });
    gScript.sendAsyncMessage("teardown");
  }

  // test main body
  async function runTests() {
    try {
      await testUpdateErrors();
      await testNoOnshippingaddresschange();
      await testNoOnshippingoptionchange();
      await testShowWithPendingPromise();
      await testEventHandlingTimeout();
      await teardown();
    } catch (error) {
      ok(false, `test_update_errors: unexpected error ${error.name}`);
      SimpleTest.finish();
    }
  }

  window.addEventListener('load', function() {
    SpecialPowers.pushPrefEnv({
      'set': [
        ['dom.payments.request.enabled', true],
        ['dom.payments.response.timeout', 1000],
      ]
    }, runTests);
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1435157">Mozilla Bug 1435157</a>
</body>
</html>
